Create vowel_and_syllable_counts.csv
This csv file accommodates a column of English words in conventional orthography; a column with the number of vowel sounds, not vowel letters, in each word; and a column reporting the number of syllables. Since a sequence comprised of a vowel followed by a sonorant consonant could either be interpreted as two syllable nuclei or a nucleus followed by a coda, some syllable counts are debatable. These values of debatable syllable counts are all ?s.